Fixed vacation service and unit test.
