+++ src/main/java/com/vacation_bot/core/vacation/IVacationService.java
@@ -1,0 +1,7 @@
+package com.vacation_bot.core.vacation;
+
+public interface IVacationService {
+
+    String createVacation(String userName, String startDate, String endDate);
+
+}
