+++ src/main/java/com/vacation_bot/gateway/inbound/VacationGateway.java
@@ -1,6 +1,6 @@
 package com.vacation_bot.gateway.inbound;
 
-import com.vacation_bot.core.vacation.VacationService;
+import com.vacation_bot.core.vacation.IVacationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
 public class VacationGateway {
 
     @Autowired
-    private VacationService vacationService;
+    private IVacationService vacationService;
 
     @PostMapping()
     public String getResponse(@RequestPart("user-name") String userName,
